+++ _controller/c_service.js
@@ -11,34 +11,46 @@ var service = {
     })
   },
   insert: function(req, res, next){
-    m_service.create(
+    if(req.body.name && req.body.unit){
+      m_service.create(
         {name: req.body.name, description: req.body.description, unit: req.body.unit})
-    .then(result =>{
-      res.send({id: result.id});
-    })
-    .catch(() =>{
-      res.status(400).send({error: 'Internal error'});
-    })
+      .then(result =>{
+        res.send({id: result.id});
+      })
+      .catch(() =>{
+        res.status(400).send({error: 'Internal error'});
+      })
+    }else{
+      res.status(404).send({error: 'It is necesary more parameters'});
+    }
   },
   edit: function(req, res, next){
-    m_service.update(        
-      {name: req.body.name, description: req.body.description, unit: req.body.unit}
-    , {where: { id: req.body.id }})
-    .then(()=>{
-      res.send({affectedRows: affectedRows});
-    })
-    .catch(() =>{
-      res.status(400).send({error: 'Internal error'});
-    })
+    if(req.body.id && req.body.name && req.body.unit){
+      m_service.update(        
+        {name: req.body.name, description: req.body.description, unit: req.body.unit}
+        , {where: { id: req.body.id }})
+      .then(result=>{
+        res.send({affectedRows: result[0]});
+      })
+      .catch(() =>{
+        res.status(400).send({error: 'Internal error'});
+     })
+    }else{
+      res.status(404).send({error: 'It is necesary more parameters'});
+    }
   },
   delete: function(req, res, next){
-    m_service.destroy({where: {id: req.body.id }})
-    .then(affectedRows =>{
-      res.send({affectedRows: affectedRows});
-    })
-    .catch(() =>{
-      res.status(400).send({error: 'Internal error'});
-    })
+    if(req.body.id){
+      m_service.destroy({where: {id: req.body.id }})
+      .then(result =>{
+        res.send({affectedRows: result});
+      })
+      .catch(() =>{
+        res.status(400).send({error: 'Internal error'});
+      })
+    }else{
+      res.status(404).send({error: 'It is necesary more parameters'});
+    }
   }
 }
